+++ R/dive_utils.R
@@ -75,3 +75,54 @@ init_dtcurve <- function(depth, time, ascent_speed = 10, way = c("OW", "WB")) {
   }
   return(dtcurve)
 }
+
+
+#' add_desat
+#' 
+#' Add the desaturation stops to the dive curve. It also check if there is no
+#' depth point higher and before the stops, to prevent accidents.
+#' 
+#' @param dtcurve a depth time curve in a data.frame with 2 columns depth and 
+#' time. Depths values are meters (positive values) and time is in minute.
+#' @param desat a desat object that contain the desaturation stops of given dive
+#' @param ascent_speed Ascent_speed in meter/minute. 10 m/min by default. 
+#' Most dive table advice to limite this speed to 20M/min maximum.
+#' @param secu security decompression stop of 3 min at 3 m. FALSE by default.
+#'
+#' @export
+add_desat <- function(dtcurve, desat, ascent_speed = 10, secu = FALSE) {
+  #### IDIOT PROOF ####
+  if (!inherits(dtcurve, 'data.frame') | any(is.na(dtcurve)) | 
+      any(colnames(dtcurve) != c('depth', 'time'))){
+    stop(paste('dtcurve must be a data.frame with 2 columns named',
+               'depth and time without any NA value'))
+  }
+  if (any(dtcurve$depth < 0) | !is.numeric(dtcurve$depth)) {
+    stop("depth must be positive numeric value(s).")
+  }
+  if (any(dtcurve$time < 0) | !is.numeric(dtcurve$time)) {
+    stop("time must be positive numeric value(s).")
+  }
+  if (any(dtcurve$time != sort(dtcurve$time))) {
+    stop("time values need to be sorted, you don't own a subaquatic dolorean")
+  }
+  
+  if(!is.desat(desat)){
+    stop("desat must be of class desat made with desat_model function")
+  }
+  
+  if (any(ascent_speed <= 0) | !is.numeric(ascent_speed) |
+      length(ascent_speed) > 1) {
+    stop("ascent_speed must be a single positive numeric value(s).")
+  }
+  if( !is.logical(secu) | is.na(secu) )
+    stop('secu must be TRUE or FALSE')
+  
+  if(is.null(desat$hour)){
+    print(1)
+  } else {
+    print(2)
+  }
+  
+  return(dtcurve)
+}